requirements.txt
glfw==2.4.0
igl==2.2.1
joblib==0.17.0
numpy==1.21.6
opencv-python==4.5.5.64
plyfile==1.0.2
PyOpenGL==3.1.0
pyrender==0.1.45
pytorch3d==0.7.4
PyYAML==6.0.1
scikit-image==0.17.2
scikit-learn==0.23.2
screeninfo==0.8.1
setuptools==50.3.1.post20201107
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tqdm==4.50.2
trimesh==3.9.15
